fix(ci): correct fast scroller height type
Use Int coerceAtLeast on IntSize.height and convert to Float for progress calculation.

Bump version to 0.2.35 (code 48).

Made-with: Cursor

+++ app/src/main/java/dev/nemeyes/ncarousel/ui/library/LibraryScreen.kt
@@ -279,7 +279,7 @@ private fun FastScroller(
                             onDragActiveChange(false)
                         },
                     ) { change, _ ->
-                        val h = size.height.coerceAtLeast(1f)
+                        val h = size.height.coerceAtLeast(1).toFloat()
                         onJumpToProgress(change.position.y / h)
                         change.consume()
                     }
